feat(nifs): impl `protogalaxy::compute_G`
refactor(nifs): mv `folded_trace` to mod

fix(util): err in doc-test

docs(nifs): `protogalaxy::poly`
