+++ src/nifs/protogalaxy/poly/mod.rs
@@ -7,15 +7,17 @@ use ff::{Field, PrimeField};
 use halo2curves::CurveAffine;
 use itertools::*;
 use num_traits::Zero;
-use rayon::prelude::*;
 use tracing::*;
 
 use crate::{
     fft,
-    plonk::{self, eval, GetChallenges, GetWitness, PlonkStructure, PlonkWitness},
+    plonk::{self, eval, GetChallenges, GetWitness, PlonkStructure},
     polynomial::{expression::QueryIndexContext, lagrange, univariate::UnivariatePoly},
     util::TryMultiProduct,
 };
+
+mod folded_trace;
+use folded_trace::FoldedTrace;
 
 #[derive(Debug, thiserror::Error, PartialEq, Eq, Clone)]
 pub enum Error {
@@ -55,6 +57,12 @@ pub enum Error {
 ///
 /// Each β here is a vector of all `X`, and each node except leaves contains all counted
 /// Each `f` here is fₙₘ =  fₙ * 1 + fₘ * βⁱ
+///
+/// # Note
+///
+/// Unlike [`compute_G`] where `X` challenge affects the nodes of the tree and generates multiple
+/// values from them, here multiple values are generated by edges, and they are stored everywhere
+/// except leaves.
 #[instrument(skip_all)]
 pub(crate) fn compute_F<C: CurveAffine>(
     beta: C::ScalarExt,
@@ -167,43 +175,40 @@ pub(crate) fn compute_F<C: CurveAffine>(
     }
 }
 
-struct FoldedTrace<F: PrimeField> {
-    witness: PlonkWitness<F>,
-    challenges: Vec<F>,
-}
-
-impl<F: PrimeField> FoldedTrace<F> {
-    fn new(
-        points_for_fft: &[F],
-        fft_domain_size: u32,
-        accumulator: impl Sync + GetChallenges<F> + GetWitness<F>,
-        traces: &[(impl Sync + GetChallenges<F> + GetWitness<F>)],
-    ) -> Box<[Self]> {
-        let folded_witnesses_collection = fold_witnesses(points_for_fft, &accumulator, traces);
-        let folded_challenges_collection =
-            fold_plonk_challenges(points_for_fft, fft_domain_size, &accumulator, traces);
-
-        folded_witnesses_collection
-            .into_iter()
-            .zip(folded_challenges_collection)
-            .map(|(witness, challenges)| Self {
-                witness,
-                challenges,
-            })
-            .collect()
-    }
-}
-impl<F: PrimeField> GetChallenges<F> for FoldedTrace<F> {
-    fn get_challenges(&self) -> &[F] {
-        &self.challenges
-    }
-}
-impl<F: PrimeField> GetWitness<F> for FoldedTrace<F> {
-    fn get_witness(&self) -> &[Vec<F>] {
-        &self.witness.W
-    }
-}
-
+/// This function calculates G(X), which mathematically looks like this:
+///
+/// $$G(X)=\sum_{i=0}^{n-1}\operatorname{pow}_i(\boldsymbol{\beta}+\alpha\cdot\boldsymbol{\delta})f_i(L_0(X)w+\sum_{j\in[k]}L_j(X)w_j)$$
+///
+/// - `f_i` - iteratively all gates for all rows sequentially. The order is taken from
+///           [`plonk::iter_evaluate_witness`].
+/// - `pow_i` - `i` degree of challenge
+/// - `L` - lagrange poly
+///
+/// # Algorithm
+///
+/// We use [`Itertools::tree_reduce`] & store in each node `X` points, for each X challenge
+///
+/// I.e. item `i` from this iterator is a collection of [pow_i(X0), pow_i(X1), ...]
+///
+/// f₀  f₁ f₂  f₃  f₄  f₅  f₆  f₇
+/// │   │  │   │   │   │   │   │
+/// 1   β  1   β   1   β   1   β
+/// │   │  │   │   │   │   │   │
+/// └───f₀₁└───f₂₃ └───f₄₅ └───f₆₇
+///     │      │       │       │
+///     1      β²      1       β²
+///     │      │       │       │
+///     └──────f₀₁₂₃   └───────f₄₅₆₇
+///            │               │
+///            └───────────────f₀₁₂₃₄₅₆₇
+///
+/// Each β here is a `beta_stroke` argument
+/// Each `f` here is vector (leafs too) of elements for each challenge with: fₙₘ =  fₙ * 1 + fₘ * βⁱ
+///
+/// # Note
+///
+/// Unlike [`compute_F`] where `X` challenge affects the edges of the tree, here the set of values
+/// is in the nodes
 pub(crate) fn compute_G<C: CurveAffine>(
     beta_stroke: C::ScalarExt,
     S: &PlonkStructure<C::ScalarExt>,
@@ -258,7 +263,7 @@ pub(crate) fn compute_G<C: CurveAffine>(
         },
     }
 
-    let evaluated = FoldedTrace::new(&points_for_fft, fft_domain_size, accumulator, traces)
+    let evaluated = FoldedTrace::new(&points_for_fft, accumulator, traces)
         .iter()
         .map(|folded_trace| plonk::iter_evaluate_witness::<C>(S, folded_trace))
         .try_multi_product()
@@ -309,117 +314,6 @@ pub(crate) fn compute_G<C: CurveAffine>(
         Some(Err(err)) => Err(err.into()),
         other => unreachable!("this case must be unreachable: {other:?}"),
     }
-}
-
-/// For each `X` we must perform the operation of sum all all matrices [`PlonkWitness`] with
-/// coefficients taken from [`lagrange::iter_eval_lagrange_polynomials_for_cyclic_group`]
-///
-/// Since the number of rows is large, we do this in one pass, counting the points for each
-/// challenge at each iteration, and laying them out in separate [`PlonkWitness`] at the end.
-fn fold_witnesses<F: PrimeField>(
-    X_challenges: &[F],
-    accumulator: &(impl GetWitness<F> + Sync),
-    witnesses: &[impl Sync + GetWitness<F>],
-) -> Vec<PlonkWitness<F>> {
-    let log_n = (witnesses.len() + 1).next_power_of_two().ilog2();
-
-    let lagrange_poly_by_challenge = X_challenges
-        .iter()
-        .map(|X| {
-            lagrange::iter_eval_lagrange_polynomials_for_cyclic_group(*X, log_n)
-                .collect::<Box<[_]>>()
-        })
-        .collect::<Box<[_]>>();
-
-    let columns_count = accumulator.get_witness().len();
-    let rows_count = accumulator.get_witness()[0].len();
-
-    let mut result_matrix_by_challenge = vec![
-        PlonkWitness {
-            W: vec![vec![F::ZERO; rows_count]; columns_count],
-        };
-        X_challenges.len()
-    ];
-
-    itertools::iproduct!(0..columns_count, 0..rows_count)
-        .map(|(col, row)| {
-            iter::once(accumulator.get_witness())
-                .chain(witnesses.iter().map(GetWitness::get_witness))
-                .zip(
-                    lagrange_poly_by_challenge
-                        .iter()
-                        .map(|m| m.iter().copied())
-                        .multi_product(),
-                )
-                .fold(
-                    vec![F::ZERO; X_challenges.len()].into_boxed_slice(),
-                    // every element of this collection - one cell for each multiplier
-                    |mut cells_by_challenge, (matrix, multiplier)| {
-                        cells_by_challenge
-                            .iter_mut()
-                            .zip(multiplier.iter())
-                            .for_each(|(res, cell)| {
-                                *res += *cell * matrix[col][row];
-                            });
-
-                        cells_by_challenge
-                    },
-                )
-        })
-        .zip(
-            // Here we take an iterator that on each iteration returns [column][row] elements for
-            // each witness for its challenge
-            //
-            // next -> vec![result[0][col][row], result[1][col][row], ... result[challenges_len][col][row]]
-            result_matrix_by_challenge
-                .iter_mut()
-                .map(|matrix| matrix.W.iter_mut().flat_map(|col| col.iter_mut()))
-                .multi_product(),
-        )
-        .par_bridge()
-        .for_each(|(elements, mut results)| {
-            results
-                .iter_mut()
-                .zip(elements.iter())
-                .for_each(|(result, cell)| **result = *cell);
-        });
-
-    result_matrix_by_challenge
-}
-
-fn fold_plonk_challenges<F: PrimeField>(
-    _X_challanges: &[F],
-    _log_n: u32,
-    _accumulator: &(impl GetChallenges<F> + Sync),
-    _witnesses: &[impl Sync + GetChallenges<F>],
-) -> Vec<Vec<F>> {
-    todo!()
-}
-
-struct MultiProduct<I: Iterator> {
-    iters: Box<[I]>,
-}
-impl<I: Iterator> Iterator for MultiProduct<I> {
-    type Item = Box<[I::Item]>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.iters.iter_mut().map(|i| i.next()).collect()
-    }
-}
-trait MultiCartesianProduct: Iterator + Sized
-where
-    <Self as Iterator>::Item: Iterator + Sized,
-{
-    fn multi_product(self) -> MultiProduct<Self::Item> {
-        MultiProduct {
-            iters: self.collect(),
-        }
-    }
-}
-
-impl<I: Iterator + Sized> MultiCartesianProduct for I where
-    <Self as Iterator>::Item: Iterator + Sized
-{
 }
 
 #[cfg(test)]
@@ -514,6 +408,39 @@ mod test {
 
         assert_ne!(
             super::compute_F::<Curve>(Field::random(&mut rnd), Field::random(&mut rnd), &S, &trace,),
+            Ok(UnivariatePoly::from_iter(
+                iter::repeat(Field::ZERO).take(16)
+            ))
+        );
+    }
+
+    #[traced_test]
+    #[test]
+    fn zero_g() {
+        let (S, trace) = poseidon_trace();
+        let mut rnd = rand::thread_rng();
+
+        assert!(
+            super::compute_G::<Curve>(Field::random(&mut rnd), &S, trace.clone(), &[trace],)
+                .unwrap()
+                .iter()
+                .all(|f| f.is_zero().into())
+        );
+    }
+
+    #[traced_test]
+    #[test]
+    fn non_zero_g() {
+        let (S, mut trace) = poseidon_trace();
+        let mut rnd = rand::thread_rng();
+        trace
+            .w
+            .W
+            .iter_mut()
+            .for_each(|row| row.iter_mut().for_each(|el| *el = Field::random(&mut rnd)));
+
+        assert_ne!(
+            super::compute_G::<Curve>(Field::random(&mut rnd), &S, trace.clone(), &[trace]),
             Ok(UnivariatePoly::from_iter(
                 iter::repeat(Field::ZERO).take(16)
             ))
